activities 4.18 - 4.22

+++ controllers/blogs.js
@@ -1,29 +1,29 @@
 const Blog = require('../model/blog')
-const User = require('../model/user')
-const checkProperties = require('../utils/checkProperties')
 const handleHTTPError = require('../utils/handleHTTPError')
 const logger = require('../utils/logger')
+const {
+  validateMissingBlogProperties,
+  validateLikesField,
+} = require('../utils/validateProperties')
 
 const getBlogs = async (_, res) => {
-  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  try {
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
 
-  return res.json(blogs).end()
+    return res.json(blogs).end()
+  } catch (error) {
+    handleHTTPError(res, 'error getting blog')
+    logger.error(`❗❗❗ Error: ${error.message}`)
+  }
 }
 
-/**
- *  Add new blog and save it to the database.
- * @param {*} req.body = { title, author, url, likes}
- * @param {*} res saved blogs
- */
 const addBlog = async (req, res) => {
   try {
-    const blog = req.body
+    const { token: user, ...blog } = req.body
 
-    const user = await User.findById(blog.userId)
+    validateMissingBlogProperties(blog)
 
-    checkProperties(blog)
-
-    const newBlog = new Blog({ ...blog, user })
+    const newBlog = new Blog({ ...blog, user: user._id })
 
     const savedBlog = await newBlog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
@@ -33,7 +33,7 @@ const addBlog = async (req, res) => {
 
     return res.json(savedBlog).end()
   } catch (error) {
-    handleHTTPError(res, 'Error adding blog')
+    handleHTTPError(res, 'error adding blog')
     logger.error(`❗❗❗ Error: ${error.message}`)
   }
 }
@@ -42,12 +42,22 @@ const deleteBlog = async (req, res) => {
   try {
     const { id } = req.params
 
+    const { token: user } = req.body
+
+    const blog = await Blog.findById(id)
+
+    if (!blog) return handleHTTPError(res, 'you cannot delete this blog', 403)
+
+    if (!(blog.user._id.toString() === user._id.toString())) {
+      return handleHTTPError(res, 'you cannot delete this blog', 403)
+    }
+
     await Blog.findByIdAndDelete(id)
 
     res.status(204).end()
   } catch (error) {
-    handleHTTPError(res, 'Error deleting blog')
-    logger.error(`❗❗❗ Error: ${error.message}`)
+    handleHTTPError(res, 'error deleting blog')
+    logger.error(`❗❗❗ Error: ${error}`)
   }
 }
 
@@ -55,16 +65,7 @@ const updateBlog = async (req, res) => {
   try {
     const { id } = req.params
 
-    let { likes } = req.body
-
-    likes = Number(likes)
-
-    if (!likes) throw new Error('field cannot be updated, invalid likes field')
-
-    likes = Math.round(likes)
-
-    if (likes < 0)
-      throw new Error('field cannot be updated, invalid likes field')
+    const likes = validateLikesField(req.body.likes)
 
     const blogInDatabase = await Blog.findById(id)
 
@@ -74,8 +75,8 @@ const updateBlog = async (req, res) => {
 
     return res.send(blogInDatabase).end()
   } catch (error) {
-    handleHTTPError(res, 'Error updating blog')
-    logger.error(error.message)
+    handleHTTPError(res, 'error updating blog')
+    logger.error(`❗❗❗ Error: ${error.message}`)
   }
 }
